requirements.txt
PySide6>=6.7.0
Pillow>=9.0.0
numpy>=1.21.0
pyqtdarktheme>=2.1.0
